Added automated Import functionality for new keys
Added functionality that new keys that appear in /keys-folder will be imported to gpg keyring if not already existing. If import fails, the user will be added to blacklisted. In the following section checking the TRX, it is being checked if user is blacklisted instead of only checking if keyfile is available.