Add functionality to guarantee unique accounts
Added functionality that the accountname AND a random number is hashed to avoid account name collisions. When two users in the same second create an account that has the same name the account name will be the same although they keys are different. To avoid that two users have the same name a 5 Byte random number will be taken out of /dev/urandom that will be hashed with the account name to ensure unique account names/adresses