Use $handover_account instead of $keylist_hash
Changed the script to use variable $handover_account and $handover_account_stamp instead of $keylist_hash and $keylist_hash_stamp. This is a bugfix. It fixes the problem that variables $keylist_hash and $keylist_hash_stamp to not match with the logged on user which causes problems when creating a transaction.