Change trigger point for ledger
Changed trigger point for ledger. Instead of building a new ledger first and pack the transaction afterwards now the transaction will first be packed and if this was successfull a new ledger will be build. If creating the archive was successfull a new ledger will be build. If not the trx is removed.