Introduced parameter delete_trx_not_indexed
Introduced parameter delete_trx_not_indexed. If that parameter is set to '1' then transactions that have not been listed within the sending users index file will be deleted instead of only being ignored. If set to '0' they will just be ignored. Setting this parameter has an effect when browsing the history. If set to '1' then transactions that have not been indexed by the sender will be deleted and so will not be listed in the history. If set to '0' these transaction will appear highlighted in red in history function.