Bugfix, Change and Security improvement
Bugfix:
Fixed bug where permissions for files have not been calculated correctly. From now on a file will be created and the permissions will be extracted from that file. 

GPG passphrase:
Furthermore the way GPG catches the passphrase has been changed back to handover a variable instead of piping it.

Security improvement:
Any archive is being checked for executables in check_archive. If any executable is found, the function will return 1 and quit.