Change in crediting coinload
Changed the way the coinload is credited to the user. Before the change you could create a account and would immediately get credited the coinload. However this leads to users creating their accounts on demand. If you would need more money you would just need to create another account. We changed this so that the coinload is credited at 00:00 UTC. So users have to wait maximum 24h before they get the coinload credited.