removed change of score for receiver
Finally removed change of score for receiver in build_ledger() again. 
This is done to avoid that accounts can get stuck/frozen by huge transaction amounts. For example when automating payments the incoming of a high volume transaction could set the score of the receiver to '0' resulting in a stuck account. To avoid any scenarios like this the change of the score shall only be influenced by the users itself, but not by other users.