webwallet UPDATE
webwallet

The webwallet is a HTML, PHP and javascript based webfrontend for the wallet.

webwallet_home.tar contains the webwallet.sh script and the webwallet folder containing the webwallet_template.html and a tmp folder.
webwallet_www-data.tar contains the wallet.php script and the index.html with logo.png for the webserver www

To run the webwallet you need to have a webserver (apache or nginx) and PHP FastCGI installed and connected to your it. The webserver itself must have a configured server that allows you to access the index.html of the www-data.