removed DGM
removed possibility to add a owner for a fungible asset (DGM) as it leads to unsolvable problems in view of timing. As the ledger is being built iteratively after any detected changes there is no way to determine which transactions have been received properly on time and which have been received afterwards. Because the next time the ledger is build the client cannot distinquish between transactions received on time. So it would be impossible to correctly distribute any amounts sent to that asset. Therefore the code representing the DGM has been removed out of the wallet and maybe moved into seperate contract logic in future.